% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Position.R
\name{crs}
\alias{crs}
\title{CRS function}
\usage{
crs(x, proj = FALSE, describe = FALSE, parse = FALSE)
}
\description{
in order to plot the data with the terra::plet function, the crs(x) function should be ran locally, could be due to a bad call of CRS in the leaflet package
See terra::crs for more information about the function.
}
\keyword{internal}
