% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{dji_init}
\alias{dji_init}
\title{Set the right environment to work with python in the background}
\usage{
dji_init(virtenv = NA)
}
\arguments{
\item{virtenv}{(character) the virtual environment you want to use, default is NA and will create a specified "r-theRmalUAV" environment.}
}
\description{
\code{\link[=dji_init]{dji_init()}} initializes a virtual environment for python using \code{\link[reticulate]{reticulate}} which is needed to run the processing for the DJI cameras
}
