% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_temp_dirp_cpp}
\alias{get_temp_dirp_cpp}
\title{Thermal Image Temperature Extraction}
\usage{
get_temp_dirp_cpp(
  filepath_image,
  image_height,
  image_width,
  object_distance,
  relative_humidity,
  emissivity,
  reflected_apparent_temperature,
  filepath_dll
)
}
\arguments{
\item{filepath_image}{Path to the R-JPEG thermal image file.}

\item{image_height}{Image height in pixels.}

\item{image_width}{Image width in pixels.}

\item{object_distance}{Distance to the target in meters (1-25m).}

\item{relative_humidity}{Relative humidity of the environment (20-100\%).}

\item{emissivity}{Emissivity of the target surface (0.10-1.00).}

\item{reflected_apparent_temperature}{Reflected apparent temperature in Celsius (-40.0 to 500.0 deg C).}

\item{filepath_dll}{Path to the dynamic link library (DLL) for processing.}
}
\value{
A numeric matrix representing the temperature values of the image.
}
\description{
Extracts temperature data from an R-JPEG thermal image using DJI's Thermal SDK.
}
