% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{T_uav_clean_persec}
\alias{T_uav_clean_persec}
\title{Keeps the best image per second}
\usage{
T_uav_clean_persec(thermal_uav, number_keep = 1, remove = FALSE)
}
\arguments{
\item{thermal_uav}{Thermal.UAV object retrieved by previous functions (e.g.: `T_uav_info()`)}

\item{number_keep}{(numerical) the number of images you want to keep per second (default = 1)}

\item{remove}{(logical) use TRUE to delete the images for your local disk space, default is FALSE where images are simple replaced in a subfolder called "Removed by img_persec"}
}
\value{
an updated Thermal.UAV object and (re)moves the least sharp images
}
\description{
`T_uav_clean_persec()` ranks all images within 1 second based on sharpness and keeps x sharpest images. This function is primarily for streaming/filming thermal cameras like the ThermalCapture. The goal is to reduce the number of images before further processing.
}
