% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{T_uav_clean_thresh}
\alias{T_uav_clean_thresh}
\title{Get sharpness threshold}
\usage{
T_uav_clean_thresh(thermal_uav)
}
\arguments{
\item{thermal_uav}{Thermal.UAV object retrieved by previous functions (e.g.: `T_uav_info()`). If Thermal.Data is empty, the raw TIFFs are loaded directly from the folder containing the tiff files.}
}
\value{
This function returns a sharpness value which can be set as threshold in the `T_uav_clean_reduc()` function
}
\description{
`T_uav_clean_thresh()` estimates sharpness by iteratively asking whether an image is sharp enough, until it is sharp enough. The goal is to set your own sharpness threshold which is used in the image reduction function.
}
