% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correct.R
\name{T_uav_corr_atm}
\alias{T_uav_corr_atm}
\title{Computes atmospheric corrections}
\usage{
T_uav_corr_atm(
  thermal_uav,
  flight_height = NA,
  T_air = NA,
  rel_hum = NA,
  T_bg = 263.55,
  emiss = 0.985
)
}
\arguments{
\item{thermal_uav}{Thermal.UAV object retrieved by previous functions (e.g.: `TheRmalUAV::T_uav_info()`)}

\item{flight_height}{(numerical) the height between the SENSOR and the OBJECT OF INTEREST. Can be a single value or a vector of the same length as the number of images. If not specified it will search for GPS altitude in the exif/meta data, or use the flight_height calculated in the `T_uav_sensor_pos` function}

\item{T_air}{(numerical) the air temperature in °C during the flight. Can be a single value, a vector of same length as the number of images or a data.frame containing temperature data ("T_air") and datetime ("datetime") in UTC ("\%d.\%m.\%Y \%H:\%M:\%OS" format). If not provided the mean temperature of the images will be taken}

\item{rel_hum}{(numerical) the relative humidity in \% during the flight. Can be a single value, a vector of same length as the number of images or a data.frame containing relative humidity data ("rel_hum") and datetime ("datetime") in UTC ("\%d.\%m.\%Y \%H:\%M:\%OS" format). If not provided a value of 50\% is assumed across the entire flight.}

\item{T_bg}{(numerical) the background temperature in Kelvin, temperature measured from the aluminium panel. If not recorded set to NA and the parameter will not be accounted for.}

\item{emiss}{(numerical) the emissivity, here by default an average for plants is given (0.985)}
}
\value{
This function returns an updated Thermal.UAV object containing Thermal.Data which is atmospherically corrected
}
\description{
`T_uav_corr_atm()` executes atmospheric corrections. Based on the amount of input different corrections will be done. The equations to estimate the transmission and Ts using background temperature are based on Heinemann et al. 2020 RS https://doi.org/10.3390/rs12071075
}
\seealso{
[`TheRmalUAV::T_uav_info()`] to get all the necessary information in the right format.
}
