% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correct.R
\name{T_uav_corr_ortho}
\alias{T_uav_corr_ortho}
\title{Atmospheric and emissivity correction for othomosaics}
\usage{
T_uav_corr_ortho(
  thermal_ortho = thermal_ortho,
  temp = "C",
  flight_height = NA,
  T_air = NA,
  rel_hum = NA,
  T_bg = 263.55,
  emiss = 0.985,
  method = NA,
  corrmap = NA,
  return_emis = NA,
  NDVI_veg = 0.8,
  NDVI_soil = 0.2,
  emiss_veg = 0.988,
  emiss_soil = 0.935
)
}
\arguments{
\item{thermal_ortho}{Map (SpatRaster) or path to geotiff file (chr) containing brightness temperatures}

\item{temp}{Indicate in which scale the `thermal_ortho` map represents the temperatures. Use "K" for Kelvin and "C" for degree Celcius.}

\item{flight_height}{(numerical) the height between the SENSOR and the OBJECT OF INTEREST expressed in meters.}

\item{T_air}{(numerical) the air temperature in °C during the flight. If NA it will use a trimmed mean of the temperatures in the `thermal_ortho` map.}

\item{rel_hum}{(numerical) the relative humidity in \% during the flight. If not provided a value of 50\% is assumed.}

\item{T_bg}{(numerical) the background temperature in Kelvin, temperature measured from the aluminium panel. If not recorded set to NA and the parameter will not be accounted for.}

\item{emiss}{(numerical) the emissivity, here by default an average for plants is given (0.985)}

\item{method}{Indicate which variable the correction map provides. Use "LC" when landcover is provided, "NDVI" if you want to use the NDVI method or "EM" when emissivity is provided. Set to NA if no spatial information on the emissivity can be calculated.}

\item{corrmap}{Map (SpatRaster) or path to geotiff file (chr) for emissivity correction. This can either be a landcover map with the right labels, an NDVI map or a emissivity map.}

\item{return_emis}{(logical) when set to TRUE, the calcualted emissivty map is returned together with the surface temperature as raster stack}

\item{NDVI_veg}{the NDVI value of vegetation. NDVI values above this threshold will receive an emissivity as defined in `emiss_veg`}

\item{NDVI_soil}{the NDVI value of bare soil. NDVI values below this threshold will receive an emissivity as defined in `emiss_soil`}

\item{emiss_veg}{the emissivity for vegetation, will be taken for an NDVI higher than `NDVI_veg`}

\item{emiss_soil}{the emissivity for bare soil, will be taken for an NDVI lower than `NDVI_soil`}
}
\value{
returns a thermal map with atmospherically corrected temperatures, and accounted for (spatially varying) emissivity.
}
\description{
`T_uav_corr_ortho()` executes atmospheric and emissivity corrections on already stitched maps.
This function can be used if only single values for the atmospheric parameters collected or when data from integrated sensors have difficulties with stitching thermal data separately.
Within flight variations are not captured in this function.
}
