% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correct.R
\name{T_uav_corr_smooth}
\alias{T_uav_corr_smooth}
\title{Smooths thermal data}
\usage{
T_uav_corr_smooth(
  thermal_uav,
  method = "image",
  T_air = NA,
  smooth_length = NA
)
}
\arguments{
\item{thermal_uav}{Thermal.UAV object retrieved by previous functions (e.g.: `T_uav_info()`). Note in order to smooth the thermal data, T_uav_corr_atm should be ran at first}

\item{method}{One of the two methods: (i) "T_air" the dataset is smoothed based on a high resolution air temperature dataset or (ii) "image" smoothing is done based on the average image temperature, default is "image"}

\item{T_air}{For method "T_air": the air temperature in °C during the flight. Can be a vector of same length as the number of images or a data.frame containing temperature data ("T_air") and datetime ("datetime") in UTC ("\%d.\%m.\%Y \%H:\%M:\%OS" format). If already provided in `T_uav_corr_atm`, you can leave it black here. Default is NA}

\item{smooth_length}{(numerical) is the amount of images that will be taken into account to smooth the temperature. Default is NA where the smooth length is the number of overlapping images.}
}
\value{
This function returns an updated Thermal.UAV object with smoothed Thermal.Data
}
\description{
`T_uav_corr_smooth()` smooths the thermal dataset to avoid large fluctuations. Two methods are possible: (i) based on a provided high temporal resolution air temperature dataset or (ii) empirical smoothing using the temperatures of the images.
}
