% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{T_uav_export}
\alias{T_uav_export}
\title{Exports thermal data as tiff-files}
\usage{
T_uav_export(thermal_uav, export_path = NA)
}
\arguments{
\item{thermal_uav}{Thermal.UAV object retrieved by previous functions (e.g.: `T_uav_info()`).}

\item{export_path}{path to the folder where the images should be stored. If NA, images are stored in a new folder in the original path called 'Ts'}
}
\value{
exports the Thermal.Data as tiff files. Note data is stored as centikelvin.
}
\description{
`T_uav_export()` exports the Thermal.Data. Note in order to export the data, at least T_uav_corr_atm should have been ran first
}
