% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{T_uav_info}
\alias{T_uav_info}
\title{Gets all the necessary info}
\usage{
T_uav_info(path, sensor = "ThermalCapture", meta_csv = NA, flight_height = NA)
}
\arguments{
\item{path}{The path to the folder containing the tiff files, or path to 1 tiff file}

\item{sensor}{(character) Indicates the sensor alias name as given in the sensor.csv file}

\item{meta_csv}{The Patch to an ddditional meta data can be provided in the form of a csv, if not provided all info will be derived from exif data (which may limit some function options)}

\item{flight_height}{(numerical) the height between the SENSOR and the GROUND. can be a single value or a vector of the same length as the number of images. If not specified it will search for GPS altitude in the exif/meta data}
}
\value{
A Thermal.UAV object with all the necessary information, serves as input in the following functions
}
\description{
`T_uav_info()` gets all the information needed for further analysis
}
