% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Position.R
\name{T_uav_pos_coreg}
\alias{T_uav_pos_coreg}
\title{Coregister locations with another camera}
\usage{
T_uav_pos_coreg(
  thermal_uav,
  opt_cameras = NA,
  rig_offset = c(0, 0, 0, 0, 0, 0),
  timediff = 0
)
}
\arguments{
\item{thermal_uav}{Thermal.UAV object retrieved by previous functions (e.g.: `T_uav_info()`).}

\item{opt_cameras}{(data.frame) containing the optimized camera positions. In the right format, can be prepared with the `T_uav_prep_coreg()` function}

\item{rig_offset}{(numerical) vector containing the different offsets of the thermal camera  compared to the reference optimized camera with following offsets in mm or ° (x, y, z, yaw, pitch, roll). Default is a vector of zeros}

\item{timediff}{(numerical) if there is a time difference (in seconds) between the optimized cameraset and the thermal data, it can be specified here, default is 0}
}
\value{
This function returns an updated Thermal.UAV object containing optimized thermal camera positions
}
\description{
`T_uav_pos_coreg()` corrects the GPS-info of the thermal data with more accurate GPS information of a coregistered (high resolution) camera. Could also be that the cameras of this coregistered sensor are already aligned in a structure form motion software.
}
