% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Position.R
\name{T_uav_pos_sensor}
\alias{T_uav_pos_sensor}
\title{Gets the information about the locations of the cameras}
\usage{
T_uav_pos_sensor(thermal_uav, extent = FALSE, overlap = FALSE, export = FALSE)
}
\arguments{
\item{thermal_uav}{Thermal.UAV object retrieved by previous functions (e.g.: `T_uav_info()`)}

\item{extent}{(logical) use TRUE to get the extents instead of the GPS locations, default is FALSE}

\item{overlap}{(logical) use TRUE if you want to calculate the average overlap between the subsequent images (only if `extent` = TRUE), default is FALSE}

\item{export}{(logical) use TRUE if you want to export the GPS/extents as a shapefile with the image names as attributes into a new folder "Shapefiles" within the path folder, default is FALSE}
}
\value{
an updated Thermal.UAV object containing additional positional information
}
\description{
`T_uav_pos_sensor()` gives the location of all images with the option to retrieve the extents overlap. Furthermore, it gives the option to export the GPS/extent information
}
