% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Position.R
\name{T_uav_pos_view}
\alias{T_uav_pos_view}
\title{Shows the cameras/image extents on an interactive map}
\usage{
T_uav_pos_view(thermal_uav, extent)
}
\arguments{
\item{thermal_uav}{Thermal.UAV object retrieved by previous functions (e.g.: `T_uav_info()`)}

\item{extent}{(logical) use TRUE if you want to plot the image extents. NOTE: only if extents are previously calculated in the `T_uav_pos_sensor()` function}
}
\value{
shows the GPS locations or the extents in an interactive way
}
\description{
`T_uav_pos_view()` shows the extents or sensor positions in the viewer as interactive map. NOTE: the function `T_uav_pos_sensor()` should have been run previously
}
