% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Position.R
\name{T_uav_prep_coreg}
\alias{T_uav_prep_coreg}
\title{Prepares data for coregistration}
\usage{
T_uav_prep_coreg(
  img_path,
  SfM_option = NA,
  opt_camera_path = NA,
  epsg = "4326",
  sensor_name = "Altum-PT_MSP",
  label = NA,
  timezone = NA
)
}
\arguments{
\item{img_path}{the folder containing the high resolution images}

\item{SfM_option}{if the cameras were first aligned in a SfM software options are: "Agisoft Metashape", "Pix4DMapper". Use NA if this was not the case and you want to run it without previous alignement. Default is NA}

\item{opt_camera_path}{path to the .txt or .csv file where the optimized camera positions is stored (only if SfM_option is not NA), defualt is NA}

\item{epsg}{the epsg in which the coordinates are exported/stored, by default "4326" (WGS 84, latitude/longitude coordinate system based on the Earth's center of mass)}

\item{sensor_name}{The name of the high resolution sensor of which the optimized GPS data will be used}

\item{label}{needed for multispectral data, it is the subscript/sequence describing the band/sensor on which everything is calibrated (in case of the Micasence Altum-PT it is the panchromatic band: "_6"). If no label enter ''}

\item{timezone}{the timezone in which the datetime data is stored. Default is NA and this preset will use the timezone of your computer system}
}
\value{
This function returns a data.frame containing positional information about optimized cameras which can be used as input in the `T_uav_pos_coreg()` function
}
\description{
`T_uav_prep_coreg()` prepares the data from a different sensor (with higher GPS accuracy) to use in the `T_uav_pos_coreg()` function.
}
