% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{T_uav_sensors}
\alias{T_uav_sensors}
\title{Provides a list of supported sensors}
\usage{
T_uav_sensors()
}
\value{
prints an overview of the available sensors
}
\description{
`T_uav_sensors()` provides an overview of all the sensors which can be used for the processing
}
