% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\docType{class}
\name{ThermalUAV-Class}
\alias{ThermalUAV-Class}
\title{Class "ThermalUAV"}
\description{
\code{\link[=ThermalUAV-Class]{ThermalUAV}} groups all the necessary information together in a logical order.
The object of Class ThermalUAV serves in most functions as input as it provides the data upon which most functions are build.
It consists of other thermal classes, each representing different kinds of information.
}
\section{Slots}{

\describe{
\item{\code{Info}}{(S4 ThermalInfo) Holds general information regaring the information of the image files and project such as path, filenames, camera info, meta data, time etc.}

\item{\code{Position}}{(S4 ThermalPosition) Holds information about the position of the tifs including gps location, flightheight, overlap as well as shapefiles (when \code{\link[=tuav_loc]{tuav_loc()}} is run)}

\item{\code{Sharpness}}{(S4 ThermalSharpness) Holds information about sharpness which is used for cleaning (e.g.: Sharpness threshold, cleaning methods, number of images removed. etc.)}

\item{\code{Atmosphere}}{(S4 ThermalAtmosphere) Holds information about the atmospheric conditions during the flight including air temperature, relative humidity, atm. transmission, emissivity}

\item{\code{Smooth}}{(S4 ThermalSmooth) Holds information about the thermal smoothing correction such as: smooth length, method etc.}

\item{\code{ThermalData}}{(list) a list of matrices holding the thermal tiff files. Depending on which functions are already run, it represents raw or corrected data.}
}}

\examples{
# Creates an empty ThermalUAV class
thermal_class <- methods::new("ThermalUAV")
}
