% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{tuav_cameras}
\alias{tuav_cameras}
\title{A list of supported cameras}
\usage{
tuav_cameras()
}
\value{
prints an overview of the available cameras
}
\description{
\code{\link[=tuav_cameras]{tuav_cameras()}} provides an overview of all the cameras which can be used for the processing
}
