% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Position.R
\name{tuav_coreg}
\alias{tuav_coreg}
\title{Co-register locations with another camera}
\usage{
tuav_coreg(
  thermal_uav,
  opt_cameras = NA,
  rig_offset = c(0, 0, 0, 0, 0, 0),
  timediff = 0
)
}
\arguments{
\item{thermal_uav}{\code{\link[=ThermalUAV-Class]{ThermalUAV}} object retrieved by previous functions (e.g.: \code{\link[=tuav_create]{tuav_create()}}).}

\item{opt_cameras}{(data.frame) containing the optimized camera positions. In the right format, can be prepared with the \code{\link[=coreg_prep]{coreg_prep()}} function}

\item{rig_offset}{(numerical) vector containing the different offsets of the thermal camera  compared to the reference optimized camera with following offsets in mm or ° (x, y, z, yaw, pitch, roll). Default is a vector of zeros}

\item{timediff}{(numerical) if there is a time difference (in seconds) between the optimized camera set and the thermal data, it can be specified here, default is 0}
}
\value{
This function returns an updated \code{\link[=ThermalUAV-Class]{ThermalUAV}} object containing optimized thermal camera positions
}
\description{
\code{\link[=tuav_coreg]{tuav_coreg()}} corrects the GPS-info of the thermal data with more accurate GPS information of a coregistered (high resolution) camera. Could also be that the cameras of this coregistered camera are already aligned in a structure form motion software.
}
\seealso{
\code{\link[=coreg_prep]{coreg_prep()}}
}
