% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correct.R
\name{tuav_correct}
\alias{tuav_correct}
\title{Computes image-level corrections}
\usage{
tuav_correct(
  thermal_uav,
  flight_height = NA,
  T_air = NA,
  rel_hum = NA,
  T_bg = NA,
  SKC = TRUE,
  emiss = 0.985
)
}
\arguments{
\item{thermal_uav}{\code{\link[=ThermalUAV-Class]{ThermalUAV}} object retrieved by previous functions (e.g.: \code{\link[=tuav_create]{tuav_create()}})}

\item{flight_height}{(numerical) the height between the camera and the OBJECT OF INTEREST. Can be a single value or a vector of the same length as the number of images. If not specified it will search for GPS altitude in the exif/meta data, or use the flight_height calculated in the \code{\link[=tuav_loc]{tuav_loc()}} function}

\item{T_air}{(numerical) the air temperature in °C during the flight. Can be a single value, a vector of same length as the number of images or a data.frame containing temperature data ("T_air") and datetime ("datetime") in UTC ("\%d.\%m.\%Y \%H:\%M:\%OS" format). If not provided the mean temperature of the images will be taken}

\item{rel_hum}{(numerical) the relative humidity in \% during the flight. Can be a single value, a vector of same length as the number of images or a data.frame containing relative humidity data ("rel_hum") and datetime ("datetime") in UTC ("\%d.\%m.\%Y \%H:\%M:\%OS" format). If not provided a value of 50\% is assumed across the entire flight.}

\item{T_bg}{(numerical) the background temperature in Kelvin, temperature measured from the aluminium panel. If not recorded set to NA and the parameter will be estimated using ("T_air"), in this case, you need to specify whether the sky was clear ("SKC" = TRUE) or overcast ("SKC" = FALSE).}

\item{SKC}{(logical) if the conditions were sky clear, set to TRUE, if there was an overcast set to FALSE (default = TRUE), only needed when T_bg is not specified.}

\item{emiss}{(numerical) the emissivity, here by default an average for plants is given (0.985)}
}
\value{
This function returns an updated \code{\link[=ThermalUAV-Class]{ThermalUAV}} object containing ThermalData which is atmospherically corrected
}
\description{
\code{\link[=tuav_correct]{tuav_correct()}} executes atmospheric corrections. Based on the amount of input different corrections will be done.
}
\seealso{
\code{\link[=tuav_create]{tuav_create()}} to get all the necessary information in the right format.
}
