% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correct.R
\name{tuav_emis}
\alias{tuav_emis}
\title{Spatial emissivity correction}
\usage{
tuav_emis(
  thermal_orig,
  thermal_uav,
  temp = "K",
  corrmap,
  method = "LC",
  write_Ts = FALSE,
  filename_Ts = NA,
  write_emiss = FALSE,
  filename_emiss = NA,
  NDVI_veg = 0.8,
  NDVI_soil = 0.2,
  emiss_veg = 0.988,
  emiss_soil = 0.935,
  LC_emiss_matrix = NA
)
}
\arguments{
\item{thermal_orig}{Map (\code{\link[terra:SpatRaster]{SpatRaster}}) or path to geotiff file (chr) to the original land surface temperature map}

\item{thermal_uav}{the \code{\link[=ThermalUAV-Class]{ThermalUAV}} object corresponding to this thermal map. Is needed to extract parameters used in \code{\link[=tuav_correct]{tuav_correct()}}}

\item{temp}{indicate in which scale the thermal map represents the temperatures. Use "K" for Kelvin and "C" for degree Celcius.}

\item{corrmap}{Map (\code{\link[terra:SpatRaster]{SpatRaster}}) or path to geotiff file (chr) for emissivity correction. This can either be a landcover map, an NDVI map or a emissivity map.}

\item{method}{Indicate which variable the correction map provides. Use "LC" when landcover is provided, "NDVI" if you want to use the NDVI method or "EM" when emissivity is provided.}

\item{write_Ts}{(logical) use TRUE if the corrected thermal map must be written. Default = FALSE}

\item{filename_Ts}{the path and filename where the corrected thermal map should be stored. If NA, it is written in the thermal path with label "_emis_corr". This variable is only used if write_Ts = TRUE}

\item{write_emiss}{(logical) use TRUE if the emissivity map must be written. Default = FALSE}

\item{filename_emiss}{the path and filename where the emissivity map should be stored. If NA, it is written in the thermal path with label "_emis". This variable is only used if write_emiss = TRUE}

\item{NDVI_veg}{the NDVI value of vegetation. NDVI values above this threshold will receive an emissivity as defined in emiss_veg}

\item{NDVI_soil}{the NDVI value of bare soil. NDVI values below this threshold will receive an emissivity as defined in emiss_soil}

\item{emiss_veg}{the emissivity for vegetation, will be taken for an NDVI higher than NDVI_veg}

\item{emiss_soil}{the emissivity for bare soil, will be taken for an NDVI lower than NDVI_soil}

\item{LC_emiss_matrix}{a two column matrix with the first column the LC values and in the second the corresponding emissivties, only used if method = "LC"}
}
\value{
returns a recalculated thermal map corrected for spatially varying different emissivities.
}
\description{
\code{\link[=tuav_emis]{tuav_emis()}} corrects a stitched thermal orthomosaic for emissivity, based on a high resolution landcover or emissivity map.
}
