% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{tuav_export}
\alias{tuav_export}
\title{Export thermal data as TIFF files}
\usage{
tuav_export(thermal_uav, export_path = NA)
}
\arguments{
\item{thermal_uav}{\code{\link[=ThermalUAV-Class]{ThermalUAV}} object retrieved by previous functions (e.g.: \code{\link[=tuav_cameras]{tuav_cameras()}}).}

\item{export_path}{path to the folder where the images should be stored. If NA, images are stored in a new folder in the original path called 'corrected'}
}
\value{
exports the ThermalData as tiff files. Note data is stored as centikelvin.
}
\description{
\code{\link[=tuav_export]{tuav_export()}} exports the ThermalData. Note in order to export the data, at least tuav_correct should have been ran first
}
\seealso{
\code{\link[=tuav_report]{tuav_report()}}
}
