% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Position.R
\name{tuav_loc}
\alias{tuav_loc}
\title{Get information on camera locations}
\usage{
tuav_loc(thermal_uav, extent = FALSE, overlap = FALSE, export = FALSE)
}
\arguments{
\item{thermal_uav}{\code{\link[=ThermalUAV-Class]{ThermalUAV}} object retrieved by previous functions (e.g.: \code{\link[=tuav_create]{tuav_create()}})}

\item{extent}{(logical) use TRUE to get the extents instead of the GPS locations, default is FALSE}

\item{overlap}{(logical) use TRUE if you want to calculate the average overlap between the subsequent images (only if extent = TRUE), default is FALSE}

\item{export}{(logical) use TRUE if you want to export the GPS/extents as a shapefile with the image names as attributes into a new folder "Shapefiles" within the path folder, default is FALSE}
}
\value{
an updated \code{\link[=ThermalUAV-Class]{ThermalUAV}} object containing additional positional information
}
\description{
\code{\link[=tuav_loc]{tuav_loc()}} gives the location of all images with the option to retrieve the extents overlap. Furthermore, it gives the option to export the GPS/extent information
}
\seealso{
\code{\link[=tuav_create]{tuav_create()}}, \code{\link[=tuav_view]{tuav_view()}}
}
