% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{tuav_persec}
\alias{tuav_persec}
\title{Keep the best image(s) per second}
\usage{
tuav_persec(thermal_uav, number_keep = 1, remove = FALSE)
}
\arguments{
\item{thermal_uav}{\code{\link[=ThermalUAV-Class]{ThermalUAV}} object retrieved by previous functions (e.g.: \code{\link[=tuav_create]{tuav_create()}})}

\item{number_keep}{(numerical) the number of images you want to keep per second (default = 1)}

\item{remove}{(logical) use TRUE to delete the images for your local disk space, default is FALSE where images are simple replaced in a subfolder called "Removed by img_persec"}
}
\value{
an updated \code{\link[=ThermalUAV-Class]{ThermalUAV}} object and (re)moves the least sharp images
}
\description{
\code{\link[=tuav_persec]{tuav_persec()}} ranks all images within 1 second based on sharpness and keeps x sharpest images. This function is primarily for streaming/filming thermal cameras like the ThermalCapture. The goal is to reduce the number of images before further processing.
}
