% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{tuav_reduc}
\alias{tuav_reduc}
\title{Clean the dataset based on overlap or sharpness}
\usage{
tuav_reduc(
  thermal_uav,
  method = "Overlap",
  min_overlap = 0.8,
  sharpness_threshold = 0.05,
  remove = FALSE
)
}
\arguments{
\item{thermal_uav}{\code{\link[=ThermalUAV-Class]{ThermalUAV}} object retrieved by previous functions (e.g.: \code{\link[=tuav_create]{tuav_create()}}). If ThermalData is empty, the raw TIFFs are loaded directly from the folder containing the tiff files.}

\item{method}{the method you want to choose for image reduction. This can be "Overlap" to reduce the dataset based on a minimal overlap (the final targetted overlap will be around `min_overlap`) or "Sharpness" to remove unsharp images (based on `sharpness_threshold`).}

\item{min_overlap}{(numerical) the targetted overlap (0-1) for the "Overlap" method or the criteria to acertain a minimal overlap in case of a large part of consecutive unsharp images. Default is 0.80}

\item{sharpness_threshold}{(numerical) the threshold to decide if an image is sharp or unsharp. Default is 0.05}

\item{remove}{(logical) use TRUE to delete the images for your local disk space. Default is FALSE where images are simple replaced in a subfolder called "Removed by img_reduc"}
}
\value{
This function returns an updated (reduced) \code{\link[=ThermalUAV-Class]{ThermalUAV}} object and (re)moves the least sharp images
}
\description{
\code{\link[=tuav_reduc]{tuav_reduc()}} reduces the thermal dataset, either based on sharpness or minimal desired overlap. NOTE: if extents are not yet calculated, this function will automatically calculate it.
}
