% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{tuav_sharp}
\alias{tuav_sharp}
\title{Estimates image sharpness}
\usage{
tuav_sharp(thermal_uav)
}
\arguments{
\item{thermal_uav}{\code{\link[=ThermalUAV-Class]{ThermalUAV}} object retrieved by previous functions (e.g.: \code{\link[=tuav_create]{tuav_create()}}). If ThermalData is empty, the raw TIFFs are loaded directly from the folder containing the tiff files.}
}
\value{
This function returns an updated \code{\link[=ThermalUAV-Class]{ThermalUAV}} object containing a vector with the sharpness for each image.
}
\description{
\code{\link[=tuav_sharp]{tuav_sharp()}} estimates sharpness of the thermal images.
}
