% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correct.R
\name{tuav_smooth}
\alias{tuav_smooth}
\title{Smooths thermal data}
\usage{
tuav_smooth(thermal_uav, method = "image", T_air = NA, smooth_length = NA)
}
\arguments{
\item{thermal_uav}{\code{\link[=ThermalUAV-Class]{ThermalUAV}} object retrieved by previous functions (e.g.: \code{\link[=tuav_create]{tuav_create()}}). Note in order to smooth the thermal data, tuav_correct should be ran at first}

\item{method}{One of the two methods: (i) "T_air" the dataset is smoothed based on a high resolution air temperature dataset or (ii) "image" smoothing is done based on the average image temperature, default is "image"}

\item{T_air}{For method "T_air": the air temperature in  deg C during the flight. Can be a vector of same length as the number of images or a data.frame containing temperature data ("T_air") and datetime ("datetime") in UTC ("\%d.\%m.\%Y \%H:\%M:\%OS" format). If already provided in \code{\link[=tuav_correct]{tuav_correct()}}, you can leave it black here. Default is NA}

\item{smooth_length}{(numerical) is the amount of images that will be taken into account to smooth the temperature. Default is NA where the smooth length is the number of overlapping images. The smooth length is ignored for method "T_air".}
}
\value{
This function returns an updated \code{\link[=ThermalUAV-Class]{ThermalUAV}} object with smoothed ThermalData
}
\description{
\code{\link[=tuav_smooth]{tuav_smooth()}} smooths the thermal dataset to avoid large fluctuations. Two methods are possible: (i) based on a provided high temporal resolution air temperature dataset or (ii) empirical smoothing using the temperatures of the images.
}
