% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Position.R
\name{tuav_view}
\alias{tuav_view}
\title{Plot cameras/image extents on an interactive map}
\usage{
tuav_view(thermal_uav, extent = FALSE)
}
\arguments{
\item{thermal_uav}{\code{\link[=ThermalUAV-Class]{ThermalUAV}} object retrieved by previous functions (e.g.: \code{\link[=tuav_create]{tuav_create()}})}

\item{extent}{(logical) use TRUE if you want to plot the image extents. NOTE: only if extents are previously calculated in the \code{\link[=tuav_loc]{tuav_loc()}} function}
}
\value{
shows the GPS locations or the extents in an interactive way
}
\description{
\code{\link[=tuav_view]{tuav_view()}} shows the extents or camera positions in the viewer as interactive map. NOTE: the function \code{\link[=tuav_loc]{tuav_loc()}} should have been run previously
}
\seealso{
\code{\link[=tuav_loc]{tuav_loc()}}
}
